% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_time.R
\name{append_time}
\alias{append_time}
\title{Append timestamp to a filepath}
\usage{
append_time(path)
}
\arguments{
\item{path}{(string) file path (existence is not checked)}
}
\value{
(string) A file path with time stamp appended in the format described
  in the description
}
\description{
format:
  filename__year-month-day_hour-minutes-seconds_timezone.extension
}
\examples{
append_time("hello/a.b")
}
