% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageProximity.R
\name{averageProximity}
\alias{averageProximity}
\title{Compute average proximity matrix on a dataset using randomForest}
\usage{
averageProximity(dataset)
}
\arguments{
\item{dataset}{A \eqn{n \times m} dataframe with valid column names}
}
\value{
A matrix with dimension \eqn{n \times n}
}
\description{
Each predictor(column) is estimated using other predictors. Proximity matrices of these models are averaged
}
\examples{
ap = averageProximity(iris)
dim(ap)
plot(hclust(ap))
}
