% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{condense}
\alias{condense}
\title{Sample points based on distance matrix}
\usage{
condense(d, radius, seed = 1, ...)
}
\arguments{
\item{d}{object of class 'dist'}

\item{radius}{(numeric) radius around a point}

\item{seed}{(positive integer) A seed. Default is 1.}

\item{...}{Arguments for \code{\link[dbscan]{frNN}}}
}
\value{
Index of condensed points
}
\description{
Sample points based on distance matrix so that distance between any pair is not less than a fixed radius. This will not preserve density, but does preserve local outliers.
}
\examples{
library("magrittr")
dataset <- rbind(
  MASS::mvrnorm(1000, mu = c(0, 0), Sigma = matrix(c(1, 0.3, 0.3, 1), 2))
  , MASS::mvrnorm(100, mu = c(10, 10), Sigma = matrix(c(1, 0.3, 0.3, 1), 2))
  , MASS::mvrnorm(10, mu = c(5, 5), Sigma = matrix(c(1, 0.3, 0.3, 1), 2))
  , MASS::mvrnorm(10, mu = c(0, 8), Sigma = matrix(c(1, 0.3, 0.3, 1), 2))
  , MASS::mvrnorm(10, mu = c(7, 0), Sigma = matrix(c(1, 0.3, 0.3, 1), 2))
  )
plot(dataset)
do <- dist(dataset)
dbscan::kNNdist(do, 4) \%>\% apply(1, median) \%>\% summary()

ci <- condense(do, 0.103)
dataset[ci, ] \%>\% plot()
dataset[condense(do, 0.68), ] \%>\% plot()
}
