% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2.R
\name{cor2}
\alias{cor2}
\title{cor2: Compute correlations of columns of a dataframe of mixed types}
\usage{
cor2(df, nproc = future::availableCores())
}
\arguments{
\item{df}{input data frame}

\item{nproc}{Number of parallel processes to use}
}
\value{
A simil/dist object.
}
\description{
Compute correlations of columns of a dataframe of mixed types.
  The dataframe is allowed to have columns of these four classes: integer,
  numeric, factor and character. The character column is considered as
  categorical variable.
}
\details{
The correlation is computed as follows:

  \itemize{

  \item integer/numeric pair: pearson correlation using `cor` function. The
  valuelies between -1 and 1.

  \item integer/numeric - factor/categorical pair: Anova is performed  and
  effect size is computed . The value lies between 0 and 1. \item
  factor/categorical pair: cramersV value is computed based on chisq test
  using `lsr::cramersV` function. The value lies between 0 and 1.

  }

  For a comprehensive implementation, use `polycor::hetcor`
}
\examples{
iris_cor <- cor2(iris, nproc = 1)
}
