% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pattern.R
\name{count_pattern}
\alias{count_pattern}
\title{count_pattern}
\usage{
count_pattern(vec, pattern, perl = TRUE)
}
\arguments{
\item{vec}{A character vector}

\item{pattern}{character string containing a regular expression.}

\item{perl}{TRUE or FALSE}
}
\value{
An integer
}
\description{
Count the number of elements of a vector that match the pattern
}
\examples{
count_pattern(c("aa", "ab", "ac"), "b")
}
