% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{counts}
\usage{
counts(vec)
}
\arguments{
\item{vec}{An atomic vector}
}
\value{
A named integer vector
}
\description{
Obtain counts of an atomic vector as a named integer vector
}
\examples{
counts(c(1,2,2,2,1,NA))
}
