% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutq.R
\name{cutq}
\alias{cutq}
\alias{quantilecut}
\title{cutq}
\usage{
cutq(vec, n = 10, ...)
}
\arguments{
\item{vec}{numeric/integer vector}

\item{n}{number of buckets (atleast two)}

\item{...}{extra named arguments passed to `cut`}
}
\value{
A factor
}
\description{
Discretize a numeric vector along quantiles
}
\details{
By passing extra arguments to `cut`, output can be styled
}
