% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_list_columns.R
\name{decode}
\alias{decode}
\title{decode base64 encoded object as a list}
\usage{
decode(x)
}
\arguments{
\item{x}{A encoded character vector}
}
\value{
A list or vector depending on the encoded object
}
\description{
encode a list with base64 encoding
}
