% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_list_columns.R
\name{decodeListColumns}
\alias{decodeListColumns}
\title{Decode the encoded list-columns in a dataframe using 'encodeListColumns' function}
\usage{
decodeListColumns(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
A dataframe where list-columns are encoded using base64 and those column names are appended by "__encoded". Those columns are decoded and "__encoded" is removed.
}
\description{
Encoded columns are detected, encoded with base64 and column names are appended with "__encoded" so that the decoder knows it. Other columns are not transformed. The 'decodeListColumns' function decodes only the columns whose names end with "__encoded"
}
