% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pca.R
\name{dist_pca}
\alias{dist_pca}
\title{Distance matrix with inter feature(variable) distances based on PCA}
\usage{
dist_pca(model, ...)
}
\arguments{
\item{model}{Object of class 'prcomp' or 'princomp'.}

\item{...}{Arguments to be passed to 'proxy::dist'}
}
\description{
Computes distance between features(variables) weighing the
  loadings proportional to corresponding eigenvalues
}
\details{
See the book: 'An introduction to applied multivariate analysis with
  R' by Everitt (et al) section 3.6 on 'Rescaling the principal components'
}
\examples{
pca_iris          <- stats::prcomp(scale(iris[,1:4]))
feature_dist_iris <- dist_pca(pca_iris)
feature_dist_iris
plot(stats::hclust(feature_dist_iris))

pca_iris2         <- stats::princomp(covmat = cor(iris[,1:4]), cor = TRUE)
feature_dist_iris2 <- dist_pca(pca_iris2)
feature_dist_iris2
all.equal(feature_dist_iris, feature_dist_iris2)
plot(stats::hclust(feature_dist_iris))
}
