% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassSampling.R
\name{downsample_majority}
\alias{downsample_majority}
\title{Downsample majority class to the size of minority class}
\usage{
downsample_majority(x, classVarName, arrangeVarName, seed = 100,
  shuffle = FALSE)
}
\arguments{
\item{x}{A dataframe}

\item{classVarName}{Name of the binary class column}

\item{arrangeVarName}{Name of the variable to arrange by}

\item{seed}{Seed chosen for sampling, defaults to 100}

\item{shuffle}{Whether to shuffle the dataframe at the end. This cannot be
TRUE when arrangeVarName is specified. Shuffle used the seed value seed + 1.}
}
\description{
Downsamples the majority class in binary classification setting
  to the size of minority class by randomly selecting majority class to match
  the number of minority class
}
\examples{
iris2         <- iris[c(1:25, 51:100), ]
iris2$Species <- factor(iris2$Species)
sampled <- downsample_majority(iris2, "Species")
table(sampled$Species)
}
