% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_measure.R
\name{dr_measure}
\alias{dr_measure}
\title{A measure of kNN of original and dimension reduced data}
\usage{
dr_measure(ldData, d, measure = "jaccard", ...)
}
\arguments{
\item{ldData}{(numeric matrix) Dimension reduced data}

\item{d}{('dist' object) Distances between points in the original space}

\item{measure}{Currently, "jaccard" is implemented}

\item{...}{Additional arguments to be passed to \code{\link[dbscan]{kNN}}}
}
\description{
The idea is to measure the effectiveness of dimension reduction
  methods by computing a measure using the nearest neighbors of a point in
  the original space and the reduced space. Currently, "jaccard" is
  implemented.
}
\details{
The metric used to compute distances on dimension reduced data is
  always euclidean.
}
\examples{
d_full  <- stats::dist(iris[,1:4])
newData <- stats::cmdscale(d_full)
newData_tsne <- Rtsne::Rtsne(d_full, is_distance = TRUE)[["Y"]]

vec <- dr_measure(newData, d_full, k = 10)
summary(vec)
plot(stats::density(vec))

vec_tsne <- dr_measure(newData_tsne, d_full, k = 10)
summary(vec_tsne)
plot(stats::density(vec_tsne))
}
