% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_list_columns.R
\name{encode}
\alias{encode}
\title{encode a list with base64 encoding at first depth}
\usage{
encode(x)
}
\arguments{
\item{x}{A listor a vector(essentially an iterable)}
}
\value{
A character vector
}
\description{
encode a list with base64 encoding
}
