% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeCharColumns.R
\name{encodeCharColumns}
\alias{encodeCharColumns}
\title{Encode character columns.}
\usage{
encodeCharColumns(df, columnNames = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{columnNames}{The columns which are to be encoded. If NULL, all character columns are detected and encoded.}
}
\value{
A dataframe
}
\description{
Encode a subset of character columns using base64 encoding.
}
\details{
Texual data in dataframes with comman seperators like ',' create problems writing to disk and reading back. This function encodes the a subset of character columns using base64 encoding where the resulting dataframe might be written to disk with ',' separator. The function 'decodeCharCols' decodes the encoded character columns are restores back the actual character columns.
}
\examples{
iris2 <- iris
iris2$Species <- as.character(iris2$Species)
iris_enc <- encodeCharColumns(iris2)
iris_enc
unique(iris_enc$Species)
iris_dec <- decodeCharColumns(iris_enc, "Species")
iris_dec
all.equal(iris_dec, iris2)
}
