% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{freq}
\alias{freq}
\title{Obtain frequencies of an atomic vector as proportions.}
\usage{
freq(vec)
}
\arguments{
\item{vec}{An atomic vector}
}
\value{
A vector of frequencies as proportions with same length as the
  vector.
}
\description{
NA as considered valid and their proportions are calculated.
}
\examples{
freq(c(1,2,2,2,1,NA))
}
