% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image2.R
\name{image2}
\alias{image2}
\title{ggplot2 based alternative for R's 'image'}
\usage{
image2(x, ...)
}
\arguments{
\item{x}{A numeric or integer matrix}

\item{...}{parameters to be passed to ggplot2::geom_raster}
}
\description{
Visualize a matrix
}
\examples{
image2( matrix(c(rnorm(1e5), runif(1e5)), 1e2) )
}
