% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_approx.R
\name{int_approx}
\alias{int_approx}
\title{int_approx}
\usage{
int_approx(vec, n)
}
\arguments{
\item{vec}{non-decreasing non-negative numeric vector}

\item{n}{A positive integer to which vector should sum to. Default value is
the rounded sum of the vector}
}
\value{
An integer vector
}
\description{
Produce an integer vector approximation of a non-increasing
  non-negative numeric vector such that it sums to an positive integer
}
\details{
The numeric vector is scaled to sum to n. Then, an integer vector is
  assigned the integer parts of the scaled vector. The leftout part of the
  sum is distributed based on the magnitude of the fractional part. This kind
  of approximation is useful when an algorithm produces ranking scores which
  have to allocated with positive integer constraints.
}
\examples{
int_approx(c(3.7, 3.2, 2.4, 1.5))
}
