% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_vector.R
\name{is_vector}
\alias{is_vector}
\title{is_vector}
\usage{
is_vector(vec)
}
\arguments{
\item{vec}{A vector}
}
\description{
Checks for a vector of one of these types: c("integer",
  "numeric", "character", "logical", "raw", "complex")
}
\examples{
is_vector(1:5)
is_vector(list(1:5))
is_vector(matrix(1:4, 2))
}
