% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_m.R
\name{jaccard_m}
\alias{jaccard_m}
\title{jaccard_m}
\usage{
jaccard_m(vec1, vec2)
}
\arguments{
\item{vec1}{A vector}

\item{vec2}{A vector}
}
\value{
A number
}
\description{
multiset jaccard for two sets A and B
}
\details{
Let n(a,A) be the number of occurences of element a in multiset A.
  definition: numerator/denominator where numerator   = sum( min( n(a,A) ,
  n(a,B) ) for every a in union(A,B)) and denominator = sum( max( n(a,A) ,
  n(a,B) ) for every a in union(A,B))
}
\examples{
set.seed(1); A = sample(letters, 100, TRUE)
set.seed(2); B = sample(letters, 100, TRUE)
jaccard_m(A, B)
}
