% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde2d_output.R
\name{kde2d_tidy}
\alias{kde2d_tidy}
\title{Convert the output of MASS::kde2d output into long format.}
\usage{
kde2d_tidy(kde2d_output)
}
\arguments{
\item{kde2d_output}{Ouput of MASS:kde2d output.}
}
\description{
Long format is helpful in ggplot2 plots.
}
\examples{
library("MASS")
temp <- MASS::kde2d(x = runif(100, 0, 1)
                    , y = runif(100, 0, 5)
                    , n = 500
                    )
str(temp)
kde2d_tidy(temp)
}
