% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc.R
\name{loc}
\alias{loc}
\title{classwise local outliers}
\usage{
loc(d, classes, r, thres = 0.4, minn = 3, minnStrategy = "include")
}
\arguments{
\item{d}{object of class 'dist'}

\item{classes}{(factor) Vector of classes}

\item{r}{Value of the fixed radius}

\item{thres}{Threshold for the fraction of observations of same class}

\item{minn}{Minimum number of neighbors in the neighborhood}

\item{minnStrategy}{When minnStrategy is 'include', the points with less than
minn neighbors are considered outliers.}
}
\value{
Vector of indexes of outliers
}
\description{
Detect local outliers when an observation of class A occurs in
  the neighborhood of class B using fixed radius neighborhood search. When
  the fraction of neighbors of the same class as the observation is less than
  the threshold, it is considered as an outlier.
}
\details{
Ideal conditions for using this is when there is no data imbalance
  and there are observations similar to eachother but from different classes.
  When minnStrategy is 'include', the global outliers (ones with less than
  minn neighbors) are included in the output.
}
