% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_usage.R
\name{mem_cur}
\alias{mem_cur}
\title{current memory usage}
\usage{
mem_cur(unit = "MB")
}
\arguments{
\item{unit}{A string, either "MB"(Default) or "GB"}
}
\value{
A number indicating memory usage in MB (or GB)
}
\description{
measure R session's current memory usage after garbage
  collection
}
\details{
Do garbage collection and check current memory usage
}
