% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_usage.R
\name{mem_current}
\alias{mem_current}
\title{mem_current}
\usage{
mem_current(unit = "MB")
}
\arguments{
\item{unit}{A string, either "MB"(Default) or "GB"}
}
\value{
A number indicating memory usage in MB (or GB)
}
\description{
measure R session's current memory usage
}
\details{
Do garbage collection and check current memory usage
}
