% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_usage.R
\name{mem_max}
\alias{mem_max}
\title{maximum memory usage}
\usage{
mem_max(unit = "MB")
}
\arguments{
\item{unit}{A string, either "MB"(Default) or "GB"}
}
\value{
A number indicating memory usage in MB (or GB)
}
\description{
measure R session's maximum memory usage after the last gc(reset
  = TRUE) call
}
\details{
Do garbage collection and check maximum memory usage
}
