% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_usage.R
\name{memory_max}
\alias{memory_max}
\title{memory_max}
\usage{
memory_max(unit = "MB")
}
\arguments{
\item{unit}{A string, either "MB"(Default) or "GB"}
}
\value{
A number indicating memory usage in MB (or GB)
}
\description{
measure R session's maximum memory usage
}
\details{
Do garbage collection and check maximum memory usage
}
