% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_approx.R
\name{na_approx}
\alias{na_approx}
\title{Linear interpolation of a vector}
\usage{
na_approx(vec)
}
\arguments{
\item{vec}{A numeric vector with NAs}
}
\value{
A numeric vector without NA
}
\description{
Linear interpolation of NAs in a vector using `zoo::na.approx`.
  The NAs at the beginning or at the end are replaced by their first non NA
  neighbour
}
\examples{
na_approx(c(NA, 2, 3, NA, 5, 6, NA))
}
