% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierIndex.R
\name{outlierIndex}
\alias{outlierIndex}
\title{Find the outlier rows in numeric matrix}
\usage{
outlierIndex(dataMatrix, doScale = TRUE, method = "knn", doPlot = TRUE,
  coef = 1.5, perplexity = 30, ...)
}
\arguments{
\item{dataMatrix}{A numeric matrix}

\item{doScale}{Whether to scale the data}

\item{method}{Only 'knn' is supported currently}

\item{doPlot}{Whether to compute TSNE and visualiza outliers}

\item{coef}{For 'boxplot.stats'}

\item{perplexity}{For 'Rtsne::Rtsne'}

\item{...}{specific inputs based on method (TODO)}
}
\value{
A list with: oulierIndex and tsne plot(ggplot2) object
}
\description{
and visualize them with TSNE
}
\details{
TODO
}
\examples{
result <- outlierIndex(as.matrix(iris[-143,1:4]), k = 3, representative = median)
result[[1]]
result[[2]]
}
