% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{Plot a few principal components}
\usage{
pcaPlot(model, rank = 3, color)
}
\arguments{
\item{model}{A 'prcomp' model}

\item{rank}{Number of first principal components to select}

\item{color}{A factor to color by}
}
\value{
Plots as a side effect
}
\description{
Plot a few principal components against each other, their densities colored by some factor
}
\examples{
pcaPlot(prcomp(iris[,1:4]), color = iris$Species)
}
