% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakmem.R
\name{peakmem}
\alias{peakmem}
\title{Compute peak memory used by a R script}
\usage{
peakmem(file, timeout = Inf)
}
\arguments{
\item{file}{Location of script}

\item{timeout}{Timeout for the process, in seconds, or as a difftime object.
If it is not finished before this, it will be killed.}
}
\value{
Maximum memory used in GB
}
\description{
The function runs the script using Rscript program in a
  different process and uses a python utility 'memusg'
  (https://github.com/jhclark/memusg/) to measure the peak memory usage by
  the prcocess and its child processes. This will be helpful in measuring the
  memory usage of code involving parallel processing. For example, code
  involving \code{\link{[parallel](mclapply)}} call. Place the 'memusg.py'
  file in '/usr/local/bin' directory and make it executable. The function is
  meant to measure peak memory utility in the order of GBs with long running
  parallel processes. For smaller processes, the overhead itself will distort
  the real usage. For smaller benchmarks, use \code{\link{[bench]{mark}}} or
  \code{\link{[peakRAM]{peakRAM}}}. For more details, its better to modify
  the python program.
}
