% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change.R
\name{percent_change}
\alias{percent_change}
\title{percent_change}
\usage{
percent_change(vec, from = vec[1])
}
\arguments{
\item{vec}{An integer or numeric vector}

\item{from}{An integer or a number}
}
\value{
A integer or numeric vector indicating percentage change
}
\description{
percentage change of a vector from something, typically the
  first element
}
\examples{
percent_change(c(2,1,3), from = 1)
}
