% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{predict}
\alias{predict}
\title{Scale a dataset using scaler model}
\usage{
predict(object, newdata, type = "scale")
}
\arguments{
\item{object}{A scaler object}

\item{newdata}{A numeric matrix or a dataframe to be scaled}

\item{type}{Either 'scale' or 'unscale'}
}
\description{
See \code{\link{scaler}}
}
