% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionCorrector.R
\name{predict.regressionCorrector}
\alias{predict.regressionCorrector}
\title{Predict method for regressionCorrector class}
\usage{
\method{predict}{regressionCorrector}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class 'regressionCorrector'}

\item{newdata}{New data to predict on}

\item{...}{Optional arguments}
}
\value{
A numeric vector
}
\description{
Predict method for regressionCorrector class
}
\seealso{
regression_corrector
}
