% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{predict.scaler}
\alias{predict.scaler}
\title{Scale a dataset using scaler model}
\usage{
\method{predict}{scaler}(object, ...)
}
\arguments{
\item{object}{A scaler object}

\item{...}{See details}
}
\description{
See \code{\link{scaler}}
}
\details{
\itemize{
\item newdata A numeric matrix or a dataframe to be scaled
\item type Either 'scale' or 'unscale'
}
}
