% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionCorrector.R
\name{print.regressionCorrector}
\alias{print.regressionCorrector}
\title{Print method for regressionCorrector class}
\usage{
\method{print}{regressionCorrector}(x, ...)
}
\arguments{
\item{x}{Object of class 'regressionCorrector'}

\item{...}{Currently, unused}
}
\value{
invisible NULL
}
\description{
Print method for regressionCorrector class
}
\details{
Prints the method of univariate fit, RMSE and PRMSE between actual
  and predicted output of fit on data (not the predict of
  regressionCorrector).
}
\seealso{
regression_corrector
}
