% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_rows.R
\name{remove_duplicate_rows}
\alias{remove_duplicate_rows}
\title{Remove duplicate rows considering all or a few columns}
\usage{
remove_duplicate_rows(x, columnNames, ...)
}
\arguments{
\item{x}{A dataframe}

\item{columnNames}{Names of the columns to use for computing duplicates. If not specified, all columns are used.}

\item{...}{Arguments to be passed to `duplicated` function}
}
\description{
Remove duplicate rows considering all or a few columns
}
\examples{
temp <- data.frame(x = c(1, 2, 3, 1), y = c(1, 3, 3, 1), z = c(1, 2, 3, 2))
remove_duplicate_rows(temp)
remove_duplicate_rows(temp, c("x", "y"))
}
