% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{rm_na}
\alias{rm_na}
\alias{remove_na}
\title{rm_na}
\usage{
rm_na(vec)
}
\arguments{
\item{vec}{A vector}
}
\value{
A vector without NA
}
\description{
remove NA from a vector and return it
}
\examples{
rm_na(c(1,NA, 2, NA))
}
