% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{rotate}
\usage{
rotate(vec, by = 1L)
}
\arguments{
\item{vec}{A vector}

\item{by}{An integer indicating number of positions. Default is 1L}
}
\value{
A vector after rotating it
}
\description{
rotate a vector by given number of positions forward or backwards
}
\examples{
rotate(1:10, 2)
rotate(letters, -5)
}
