% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_frac_matrix}
\alias{sample_frac_matrix}
\title{dplyr like sample_frac for a matrix}
\usage{
sample_frac_matrix(mat, frac, replace = FALSE, weight = NULL)
}
\arguments{
\item{mat}{A matrix}

\item{frac}{fraction of rows to sample}

\item{replace}{See arguments for 'base::sample'}

\item{weight}{See arguments for 'base::sample'}
}
\value{
A matrix
}
\description{
dplyr like sample_frac for a matrix
}
\examples{
library("magrittr")
mat <- matrix(1:1e4, ncol = 10)
mat \%>\% sample_frac_matrix(0.1) \%>\% dim()
}
