% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_large.R
\name{sample_large}
\alias{sample_large}
\title{Sample n things out if 1:maxValue with replacement}
\usage{
sample_large(n, maxValue, seed)
}
\arguments{
\item{n}{Number of samples expected}

\item{maxValue}{Length of the vector to sample from}

\item{seed}{Value of seed}
}
\value{
An numeric vector, with replacement sample of 1:maxValue
}
\description{
Sample from 1:maxValue vector without creating it, using runif
}
\examples{
as.character(try(sample(1:1e20, 100), silent = TRUE))
sample_large(100, 1e20)
}
