% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_n_matrix}
\alias{sample_n_matrix}
\title{dplyr like sample_n for a matrix}
\usage{
sample_n_matrix(mat, n, replace = FALSE, weight = NULL)
}
\arguments{
\item{mat}{A matrix}

\item{n}{Number of rows to sample}

\item{replace}{See arguments for 'base::sample'}

\item{weight}{See arguments for 'base::sample'}
}
\value{
A matrix
library("magrittr")
mat <- matrix(1:1e4, ncol = 10)
mat %>% sample_n_matrix(100) %>% dim()
}
\description{
dplyr like sample_n for a matrix
}
