% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler_}
\alias{scaler_}
\title{Constructor for scaler class}
\usage{
scaler_(ncols, center_vector, scale_vector)
}
\arguments{
\item{ncols}{(positive integer) Number of columns}

\item{center_vector}{Vector to be used for centering. Length should be equal
to ncols}

\item{scale_vector}{Vector to be used for scaling. Length should be equal to
ncols}
}
\description{
See \code{\link{scaler}}
}
