% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidekicks.R
\docType{package}
\name{sidekicks-package}
\alias{sidekicks}
\alias{sidekicks-package}
\title{sidekicks}
\description{
A misc set of functions used often during data analysis
}
\author{
\strong{Maintainer}: KS Srikanth \email{sri.teach@gmail.com}

}
