% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toString_to_character.R
\name{toString_to_character}
\alias{toString_to_character}
\title{Split a long delimted string into individual strings.}
\usage{
toString_to_character(string, sep = ",", stripWhitespace = TRUE)
}
\arguments{
\item{string}{Long string with delimters, may be output of 'toString'.}

\item{sep}{delimiter. Default is ",".}

\item{stripWhitespace}{Whether to trim whitespace on both sides of resulting strings.}
}
\value{
A string.
}
\description{
Usually helpful in setting extracted levels.
}
\examples{
concat <- toString(colnames(iris))
print(concat)
toString_to_character(concat)
}
