% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toString_to_character.R
\name{toString_to_characterString}
\alias{toString_to_characterString}
\title{Split a long delimted string into individual strings and use then to construct a character vector.}
\usage{
toString_to_characterString(string, sep = ",", stripWhitespace = TRUE)
}
\arguments{
\item{string}{Long string with delimters, may be output of 'toString'.}

\item{sep}{delimiter. Default is ",".}

\item{stripWhitespace}{Whether to trim whitespace on both sides of resulting strings.}
}
\value{
A string ready to constructed.
}
\description{
Usually helpful in setting extracting levels elsewhere in the code.
}
\examples{
concat <- toString(colnames(iris))
print(concat)
toString_to_characterString(concat)
eval(parse(text = toString_to_characterString(concat)))
}
