% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_fraction.R
\name{unique_fraction}
\alias{unique_fraction}
\alias{fraction_unique}
\title{unique_fraction}
\usage{
unique_fraction(vec, na.rm = TRUE)
}
\arguments{
\item{vec}{A vector}

\item{na.rm}{Whether to consider NAs as a distinct element}
}
\value{
A numeric between 0 and 1
}
\description{
fraction of unique values in a vector
}
\examples{
unique_fraction(c(1, 1, 2, 2, 2))
unique_fraction(c(1, 1, 2, 2, NA, 2, NA))
unique_fraction(c(1, 1, 2, 2, NA, 2, NA), na.rm = FALSE)
}
