% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassSampling.R
\name{upsample_minority}
\alias{upsample_minority}
\title{Upsample minority class to the size of majority class}
\usage{
upsample_minority(x, classVarName, arrangeVarName, seed = 100,
  shuffle = FALSE)
}
\arguments{
\item{x}{A dataframe}

\item{classVarName}{Name of the binary class column}

\item{arrangeVarName}{Name of the variable to arrange by}

\item{seed}{Seed chosen for sampling, defaults to 100}

\item{shuffle}{Whether to shuffle the dataframe at the end. This cannot be
TRUE when arrangeVarName is specified. Shuffle used the seed value seed +
1.}
}
\description{
Upsamples the minority class in binary classification setting to
  the size of majority class by randomly selecting minority class with
  repition to match the number of majority class
}
\examples{
iris2         <- iris[c(1:25, 51:100), ]
iris2$Species <- factor(iris2$Species)
sampled <- upsample_minority(iris2, "Species")
table(sampled$Species)
}
