% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_tibble.R
\name{vector_to_tibble}
\alias{vector_to_tibble}
\title{vector_to_tibble}
\usage{
vector_to_tibble(vec, colname = "col")
}
\arguments{
\item{vec}{A vector}

\item{colname}{Name of the column to be formed}
}
\value{
A tibble
}
\description{
Utility to convert named vector to two column tibble with first column being names of the vector and second being the vector
}
\details{
The column containing names will be named 'names_colname'
}
\examples{
vector_to_tibble(c(a = 1, b = 2), colname = "val")
}
