% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierIndex.R
\name{outlierIndex}
\alias{outlierIndex}
\title{Find the outlier rows in a distance matrix}
\usage{
outlierIndex(distObject, method = "knn", doPlot = TRUE, coef = 1.5,
  perplexity = 30L, ...)
}
\arguments{
\item{distObject}{A disttance object.}

\item{method}{Only 'knn' is supported currently}

\item{doPlot}{Whether to compute TSNE and visualize outliers}

\item{coef}{For 'boxplot.stats'}

\item{perplexity}{param for Rtsne::Rtsne.}

\item{...}{specific inputs based on method}
}
\value{
A list with: oulierIndex and tsne plot(ggplot2) object
}
\description{
and visualize them with TSNE
}
\details{
For method 'knn', provide these two inputs:
\itemize{
\item k: Number of neighbors
\item representative: function to summarize the vector of distances from a point
}
}
\examples{
result <- outlierIndex(dist(iris[,1:4]), k = 3, representative = median)
result[[1]]
result[[2]]
}
