% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{replace_na}
\alias{replace_na}
\title{replace_na}
\usage{
replace_na(object, by)
}
\arguments{
\item{object}{A vector or matrix or list or dataframe or array}

\item{by}{Something appropriate to replace by}
}
\value{
Object after NA are replaced
}
\description{
Replace NA by something appropriate
}
\examples{
replace_na(c(1, 2, NA, 4), 3)
}
